requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
jinja2
